astronomers/astronomer_3/requirements.txt
yfinance==0.1.55
ipykernel==5.3.4
scikit-learn==0.23.1
papermill==2.1.2
apache-airflow-providers-papermill==1.0.0